+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Tagger Link
   - Tagger button is now positioned correctly when sorting cards by spoiled date
+- Scryfall Search
+  - Includes an inline search bar in search dialog
 
 # 0.5.0 - Pre-release